Moved style.rs from homotopy-graphics into homotopy-model to fix import errors.
