+++ homotopy-model/Cargo.toml
@@ -16,6 +16,8 @@ paste = "1.0.8"
 rmp-serde = "0.15.5"
 serde_json = "1.0.85"
 
+palette = { git = "https://github.com/Ogeon/palette", features = ["serializing"] }
+
 log = "0.4.16"
 
 thiserror = "1.0.32"
