+++ homotopy-model/src/graphics/style.rs
