+++ homotopy-model/src/lib.rs
@@ -1,3 +1,4 @@
 
 pub mod model;
+pub mod graphics;
