+++ homotopy-model/src/model/proof/generators.rs
@@ -1,5 +1,5 @@
 use homotopy_core::{common::Generator, Diagram};
-use homotopy_graphics::style::{Color, GeneratorStyle, SignatureStyleData, VertexShape};
+use crate::graphics::style::{Color, GeneratorStyle, SignatureStyleData, VertexShape};
 use serde::Serialize;
 
 use super::signature::Signature;
